% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{betadiversity_core}
\alias{betadiversity_core}
\title{Compute the beta diversity between all sites}
\usage{
betadiversity_core(mat_pa)
}
\arguments{
\item{mat_pa}{presence absence matrix (sites as rows and species as columns).}
}
\value{
A dataframe with all the combinaisons of site and the counts.
}
\description{
Compute the beta diversity between all sites for a specific presence-absence
matrix.
}
