% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_as_pa.R
\name{ec_as_pa}
\alias{ec_as_pa}
\alias{ec_as_pa.pa}
\alias{ec_as_pa.matrix}
\title{Convert foreign object to a pa object}
\usage{
ec_as_pa(x, ...)

\method{ec_as_pa}{pa}(x, ...)

\method{ec_as_pa}{matrix}(x, ...)
}
\arguments{
\item{x}{object to be converted into an object class \code{pa}.}

\item{...}{further arguments passed on to \code{\link[=ec_pa]{ec_pa()}}.}
}
\description{
Convert foreign object to a \code{pa} object.
}
