% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_betadiversity.R
\name{ec_betadiversity}
\alias{ec_betadiversity}
\title{Compute the beta diversity between all sites}
\usage{
ec_betadiversity(mat_pa, methods = "bc", site_names = NULL)
}
\arguments{
\item{mat_pa}{presence absence matrix (sites as rows and species as columns).}

\item{methods}{a vector of two-letters strings describing the methods te be used.
Values should be taken among \code{ra}, \code{bc}, \code{wi} (see details).}

\item{site_names}{string vector giving the names of the sites. If \code{NULL} a numerical sequence is used.}
}
\value{
A matrix with all the combinaison of site and the associated betadiv.
}
\description{
Compute the beta diversity between all sites for a specific
presence-absence matrix.
}
\details{
Currently \code{ra} stands fpr raw and return the number of occurrence for
the for scenarion (A, B, C, D). \code{bc} Bray-Curtis and \code{wi}
Wishart.
}
\examples{
mat <- matrix(stats::runif(20)>.5, 10)
ec_betadiversity(mat)

}
\references{
Legendre, Pierre, and Miquel De Cáceres. “Beta Diversity as the Variance of Community Data: Dissimilarity Coefficients and Partitioning.” Ecology Letters 16, no. 8 (August 2013): 951–63. https://doi.org/10.1111/ele.12141.
Koleff, P., Gaston, K. J. & Lennon, J. J. Measuring beta diversity for presence-absence data. J. Anim. Ecol. 72, 367\u2013382 (2003).
}
\author{
Kevin Cazelles
}
