% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_cooccurrence.R
\name{ec_cooc}
\alias{ec_cooc}
\alias{ec_cooc.pa}
\alias{ec_cooc.cooc_count}
\alias{ec_cooc_pairwise}
\alias{ec_cooc_pairwise.pa}
\alias{ec_cooc_pairwise.cooc_count}
\alias{ec_cooc_species}
\alias{ec_cooc_species.pa}
\alias{ec_cooc_species.cooc_count}
\alias{ec_cooc_global}
\alias{ec_cooc_global.pa}
\alias{ec_cooc_global.cooc_count}
\title{Cooccurrence analysis}
\usage{
ec_cooc(x, ...)

\method{ec_cooc}{pa}(x, ...)

\method{ec_cooc}{cooc_count}(x, nsit, ...)

ec_cooc_pairwise(x, ...)

\method{ec_cooc_pairwise}{pa}(x, ...)

\method{ec_cooc_pairwise}{cooc_count}(x, nsit, ...)

ec_cooc_species(x, ...)

\method{ec_cooc_species}{pa}(x, ...)

\method{ec_cooc_species}{cooc_count}(x, nsit, ...)

ec_cooc_global(x, ...)

\method{ec_cooc_global}{pa}(x, ...)

\method{ec_cooc_global}{cooc_count}(x, nsit, ...)
}
\arguments{
\item{x}{object to be converted into an object class \code{pa}.}

\item{...}{ignored.}

\item{nsit}{number of sites.}
}
\description{
Cooccurrence analysis.
}
\section{Functions}{
\itemize{
\item \code{ec_cooc_pairwise}: Return a matrix of cooccurrence metrics for pair of species.

\item \code{ec_cooc_species}: Return a matrix of cooccurrence metrics for species.

\item \code{ec_cooc_global}: Return a matrix of cooccurrence metrics for species.
}}

\examples{
mat0 <- matrix(0, 10, 10)
mat1 <- matrix(1, 10, 10)
matU <- rbind(cbind(mat1, mat0), cbind(mat0, mat1))
ec_cooc(ec_as_pa(matU))
}
