% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_cooccurrence.R
\name{ec_cooc_count_pair}
\alias{ec_cooc_count_pair}
\alias{ec_cooc_count_triplet}
\alias{ec_checkerboard}
\title{Co-occurrence network}
\usage{
ec_cooc_count_pair(x)

ec_cooc_count_triplet(x)

ec_checkerboard(x)
}
\arguments{
\item{x}{a \code{pa} object or an R object to a coerced to one (see \code{\link[=ec_as_pa]{ec_as_pa()}}).}
}
\description{
Compute the co-occurrence between all pairs of species and descriptive metrics of the co-occurrence network.
}
\details{
Currently \code{bi} tests the presence of a significant value of occurrence
assuming species occurrence are independent binomial distribution. \code{by}
takes the limited number of sites into account by using an hypergeometric
distribution (see Veech 2013). Note that if the number of sites is large and
the occurrence of both species relatively low, then \code{bi} and \code{hy}
give very similar results.
}
\section{Functions}{
\itemize{
\item \code{ec_cooc_count_pair()}: A matrix with all pairs of species and the corresponding co-occurrence counts.

\item \code{ec_cooc_count_triplet()}: A matrix with all triplets of species and the corresponding co-occurrence counts.

\item \code{ec_checkerboard()}: Compute the checkerboard score and return a list of three elements:
\itemize{
\item \code{units} which incudes checkerboard units and t
\item \code{c_score} checkerboard scores.
\item \code{c_score_s2} the S2 statistics in Roberts & Stone (1990).
}

}}
\examples{
mat <- ec_generate_pa(1000, 6, .2)
out <- ec_cooc_count_pair(mat)
#plot(out$zs_bi*sqrt(1/0.2), out$zs_hy)
#abline(0,1)

# Classical example, in Stone & Roberts 1990
mat0 <- matrix(0, 10, 10)
mat1 <- matrix(1, 10, 10)
matU <- rbind(cbind(mat1, mat0), cbind(mat0, mat1))
ec_checkerboard(matU)
}
\references{
\itemize{
\item Veech, J. A. (2013). A probabilistic model for analysing species co-occurrence: Probabilistic model. Global Ecology and Biogeography, 22(2), 252–260.
\item Arita, H. T. (2016). Species co-occurrence analysis: Pairwise versus matrix-level approaches: Correspondence. Global Ecology and Biogeography, 25(11), 1397–1400.
}

\itemize{
\item Stone, L., & Roberts, A. (1990). The checkerboard score and species distributions. Oecologia, 85(1), 74–79. https://doi.org/10.1007/BF00317345
\item Roberts, A., & Stone, L. (1990). Island-sharing by archipelago species. Oecologia, 83(4), 560–567. https://doi.org/10.1007/BF00317210
}
}
