% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_generate.R
\name{ec_generate}
\alias{ec_generate}
\title{Generate random presence/absence matrix}
\usage{
ec_generate(nst, nsp, probs)
}
\arguments{
\item{nst}{site number (determine the number of rows).}

\item{nsp}{number of species (determine the number of columns).}

\item{probs}{numeric vector of probabilities presence of species. If there are less elements than species, then the vector is recycled and if there are more, the vector is truncated (see \code{\link[=rep_len]{rep_len()}})}
}
\description{
Generate random presence/absence matrix
}
