% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_pa.R
\name{ec_pa}
\alias{ec_pa}
\title{Create a presence/absence object
Create a presence/absence (\code{pa}) object from a matrix.}
\usage{
ec_pa(x, threshold = 0, spc_name = NULL, sit_name = NULL)
}
\arguments{
\item{x}{a R object to be coerced as a matrix.}

\item{threshold}{threshold value. Above (strictly) this value, a species is considered present.}

\item{spc_name}{vector of species names.}

\item{sit_name}{vector of site names.}
}
\description{
Create a presence/absence object
Create a presence/absence (\code{pa}) object from a matrix.
}
\details{
If there are less names than species (or sites), or no names at all, then names will be automatically added. These names are created using column or row numbers (zero-padded to the number of digits of total the number of column or row), preceded by \code{spc_} for column and \code{sit_} for sites.
}
\examples{
ec_pa(matrix(c(0,0,2,1,1,0), 3, 2), spc_name = "Lynx", sit_name = "GHQ") 
}
