% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_rarefaction.R
\name{ec_rarefaction}
\alias{ec_rarefaction}
\title{Compute a rarefaction curve}
\usage{
ec_rarefaction(matpa, nrep = 100)
}
\arguments{
\item{matpa}{presence absence matrix (sites as rows and species as columns).}

\item{nrep}{number of replicates.}
}
\value{
A matrix with rarefaction replicates as rows.
}
\description{
Compute a rarefaction curve based on a presence-basence matrix.
}
\references{
Gotelli, Nicholas J., and Robert K. Colwell. Quantifying Biodiversity: Procedures and Pitfalls in the Measurement and Comparison of Species Richness. Ecology Letters (2001).
}
\author{
Kevin Cazelles
}
