% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_temporal_betadiversity.R
\name{ec_temporal_betadiversity}
\alias{ec_temporal_betadiversity}
\title{Compute temporal betadiversity}
\usage{
ec_temporal_betadiversity(mat_pa1, mat_pa2, methods = "bc",
  site_names = NULL)
}
\arguments{
\item{mat_pa1}{a presence-absence matrix at time 1.}

\item{mat_pa2}{a presence-absence matrix at time 2.}

\item{methods}{a vector of two-letters strings describing the methods te be used.
Vaues should be taken among \code{ra}, \code{bc}, \code{wi} (see details).}

\item{site_names}{string vector giving the names of the sites.}
}
\value{
A data frame with desired output as columns.
}
\description{
Compute temporal betadiversity.
}
\examples{
ec_temporal_betadiversity(rbind(c(1, 0), c(1, 1)), rbind(c(0, 1), c(1, 1)))

}
\author{
Kevin Cazelles
}
