% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_as_pa.R
\name{print.pa}
\alias{print.pa}
\title{Print method for pa object}
\usage{
\method{print}{pa}(x, ..., max_sit = 10, max_spc = max_sit)
}
\arguments{
\item{x}{object of \code{pa}.}

\item{...}{ignored.}

\item{max_sit}{maximum number of sites to be printed.}

\item{max_spc}{maximum number of species to be printed.}
}
\description{
Print method for pa object.
}
