% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rarefaction_core}
\alias{rarefaction_core}
\title{Compute rarefaction}
\arguments{
\item{mat_pa}{presence absence matrix (sites as rows and species as columns).}

\item{nrep}{an integer specifying the number of repetition.}
}
\description{
Compute the rarefaction curve for a given presence-absence matrix.
}
