% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{temporal_beta_core}
\alias{temporal_beta_core}
\title{Compute the beta diversity two set of sites}
\usage{
temporal_beta_core(mat_pa1, mat_pa2)
}
\arguments{
\item{mat_pa1}{presence absence matrix (sites as rows and species as columns) for time 1.}

\item{mat_pa2}{presence absence matrix for time 2.}
}
\value{
A dataframe with all the combinaison of sites and the counts associated.
}
\description{
Compute the beta diversity between all sites for a specific presence-absence
matrix.
}
\author{
Kevin Cazelles
}
