% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_betadiversity.R
\name{ec_betadiversity}
\alias{ec_betadiversity}
\title{Compute beta diversity}
\usage{
ec_betadiversity(mat_pa, methods = "bc", site_names = NULL)
}
\arguments{
\item{mat_pa}{presence absence matrix (sites as rows and species as columns).}

\item{methods}{a vector of two-letters strings describing the methods te be used.
Values should be taken among \code{ra}, \code{bc}, \code{wi} and \code{ja} (see details).}

\item{site_names}{string vector giving the names of the sites. If \code{NULL} a numerical sequence is used.}
}
\value{
A matrix with all the combinaison of site and the associated betadiversity.
}
\description{
Compute the beta diversity between all pair of sites for a specific
presence-absence matrix.
}
\details{
Currently \code{ra} stands for raw and returns the number of occurrence.
Additional values are
\itemize{
\item \code{bc}: Bray-Curtis index,
\item \code{wi}: Wishart index,
\item \code{ja}: Jaccard index.
}
}
\examples{
mat <- ec_generate(20, 10, .4)
ec_betadiversity(mat)

}
\references{
\itemize{
\item Legendre, P., and De Caceres M.. Beta Diversity as the Variance of Community
Data: Dissimilarity Coefficients and Partitioning. Ecology Letters (2013).
\item Koleff, P., Gaston, K. J. & Lennon, J. J. Measuring beta diversity for presence-absence data. Journal of Animal Ecology (2003).
}
}
