% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_cooccurrence.R
\name{ec_cooccurrence}
\alias{ec_cooccurrence}
\title{Compute the co-occurrence between all pairs of species}
\usage{
ec_cooccurrence(mat_pa, species_names = NULL)
}
\arguments{
\item{mat_pa}{presence absence matrix (sites as rows and species as columns).}

\item{species_names}{string vector giving the names of the sites. If}
}
\value{
A matrix with species combination and co-occurrence values associated.
}
\description{
Compute the co-occurrence between all pair of species.
}
\examples{
mat <- matrix(stats::runif(30)>.5, 10)
ec_cooccurrence(mat)
}
\author{
Kevin Cazelles
}
