% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_cooccurrence.R
\name{ec_cooccurrence}
\alias{ec_cooccurrence}
\alias{ec_overlap}
\title{Co-occurrence network}
\usage{
ec_cooccurrence(mat_pa, test = NULL, species_names = NULL)

ec_overlap(mat_pa)
}
\arguments{
\item{mat_pa}{presence absence matrix (sites as rows and species as columns).}

\item{test}{test to be performed. Currenlty \code{bi} and \code{hy}
are available (see details). Default is set to \code{NULL}, meaning
no test is performed.}

\item{species_names}{string vector giving the names of the sites. If}
}
\description{
Compute the co-occurrence between all pairs of species and descriptive metrics of the co-occurrence network.
}
\details{
Currently \code{bi} tests the presence of a significant value of occurrence
assuming species occurrence are independent binomial distribution. \code{by}
takes the limited number of sites into account by using an hypergeometric
distribution (see Veech 2013). Note that if the number of sites is large and
the occurrence of both species relatively low, then \code{bi} and \code{hy}
give very similar results.
}
\section{Functions}{
\itemize{
\item \code{ec_cooccurrence}: A matrix with species combination and co-occurrence values associated.

\item \code{ec_overlap}: Return a list of two elements:
\itemize{
\item \code{overlap} which incudes overlaps for all pairs of species and the degree of symmetry (as described in Araujo 2011)
\item \code{network_contribution} that includes the species contributions to network robustness, \code{robustness}, and the species sensitivity to the loss of links in the network, \code{sensitivity} (see Araujo 2011).
}
}}

\examples{
mat <- ec_generate(1000, 6, .2)
out <- ec_cooccurrence(mat, test = c('bi', 'hy'))
plot(out$zs_bi*sqrt(1/0.2), out$zs_hy)
abline(0,1)
}
\references{
\itemize{
\item Veech, J. A.. A Probabilistic Model for Analysing Species Co-Occurrence:
Probabilistic Model. Global Ecology and Biogeography (2013).
}

Araujo, M. B., Rozenfeld, A., Rahbek, C., & Marquet, P. A. (2011). Using species co-occurrence networks to assess the impacts of climate change. Ecography, 34(6), 897–908. https://doi.org/10.1111/j.1600-0587.2011.06919.x
\itemize{
\item Griffith, G. P., Strutton, P. G., & Semmens, J. M. (2018). Climate change alters stability and species potential interactions in a large marine ecosystem. Global Change Biology, 24(1), e90–e100. https://doi.org/10.1111/gcb.13891
}
}
