% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_cooccurrence.R
\encoding{latin1}
\name{ec_cooccurrence}
\alias{ec_cooccurrence}
\title{Co-occurrence}
\usage{
ec_cooccurrence(mat_pa, test = NULL, species_names = NULL)
}
\arguments{
\item{mat_pa}{presence absence matrix (sites as rows and species as columns).}

\item{test}{test to be performed. Currenlty \code{bi} and \code{hy}
are available (see details). Default is set to \code{NULL}, meaning
no test is performed.}

\item{species_names}{string vector giving the names of the sites. If}
}
\value{
A matrix with species combination and co-occurrence values associated.
}
\description{
Compute the co-occurrence between all pairs of species.
}
\details{
Currently \code{bi} tests the presence of a significant value of occurrence
assuming species occurrence are independant binomial distribution. \code{by}
takes the limited number of sites into account by using an hypergeometric
distribution (see Veech 2013). Note that if the number of sites is large and
the occurrence of bith species relatively low, then \code{bi} and \code{hy}
give very similar results.
}
\examples{
mat <- matrix(stats::runif(60000)>.2,  ncol = 6)
out <- ec_cooccurrence(mat, test = c('bi', 'hy'))
plot(out$zs_bi*sqrt(1/0.2), out$zs_hy)
abline(0,1)
}
\references{
Veech, J. A.. A Probabilistic Model for Analysing Species Co-Occurrence:
Probabilistic Model. Global Ecology and Biogeography (2013).
}
\author{
Kevin Cazelles
}
