% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_rarefaction.R
\name{ec_rarefaction}
\alias{ec_rarefaction}
\title{Compute a rarefaction curve}
\usage{
ec_rarefaction(x, nrep = 100)
}
\arguments{
\item{x}{presence absence matrix (sites as rows and species as columns).}

\item{nrep}{number of replicates.}
}
\value{
A matrix with rarefaction replicates as rows.
}
\description{
Compute a rarefaction curve based on a presence-absence matrix.
}
\examples{
mat <- ec_generate(20, 10, .4)
res <- ec_rarefaction(mat, 1000)
plot(apply(res, 1, mean))
}
\references{
Gotelli, N. J., and Robert K. C.. Quantifying Biodiversity: Procedures and Pitfalls in the Measurement and Comparison of Species Richness. Ecology Letters (2001).
}
